Updated test application in order to support full custom view in Tooltip view
