+++ app/src/main/res/layout/custom_view.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<RelativeLayout
+	android:layout_width="200dp"
+	android:layout_height="300dp"
+	xmlns:android="http://schemas.android.com/apk/res/android"
+	android:background="@drawable/toast_frame">
+
+	<ImageView
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:id="@+id/button"
+		android:src="@mipmap/ic_launcher"
+		android:layout_centerInParent="true"
+		/>
+
+</RelativeLayout>
