+++ app/src/main/res/values/strings.xml
@@ -6,5 +6,6 @@
     <string name="action_settings">Settings</string>
     <string name="title_activity_main_activity2">Target Tooltip 2</string>
     <string name="title_activity_main_activity3">Target Tooltip 3</string>
+    <string name="title_activity_main_activity4">Target Tooltip 4</string>
 
 </resources>
